service-price计价计算完成

+++ online-taxi-public/internal-common/src/main/java/com/mashibing/internalcommon/dto/PriceRule.java
@@ -11,7 +11,7 @@ public class PriceRule {
 
     private double startFare;
 
-    private int startMile;
+    private double startMile;
 
     private double unitPricePerMile;
